requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
cloudpickle==3.1.1 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
farama-notifications==0.0.4 ; python_version >= "3.10" and python_version < "4.0"
gymnasium==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
minigrid==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.2.4 ; python_version >= "3.10" and python_version < "4.0"
ollama==0.4.7 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.33.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.11.1 ; python_version >= "3.10" and python_version < "4.0"
pygame==2.6.1 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.13.0 ; python_version >= "3.10" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
